requirements_dev.txt
autopep8==1.6.0
flake8==4.0.1
ipykernel==5.3.0
pylint==2.14.5
pytest==7.1.2
pytest-cov==2.9.0
pytest-django==4.5.2
pytest-flake8==1.1.1
pytest-xdist==2.5.0
python-dotenv==0.21.0
rope==0.17.0
